Edit-Delete Functionality Working

+++ src/main/java/com/prodapt/learningspring/entity/Post.java
@@ -10,7 +10,6 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
 import jakarta.persistence.PrePersist;
 import jakarta.persistence.Temporal;
 import jakarta.persistence.TemporalType;
@@ -21,7 +20,7 @@ import lombok.Data;
 public class Post {
   @Id
   @GeneratedValue(strategy=GenerationType.IDENTITY)
-  private int id;
+  private long id;
   
   private String content;
